email_generation/requirements.txt
ddgs
requests
beautifulsoup4
python-dotenv
openai